Moved monitor and dumper
