+++ atcoder-problems-backend/src/lib.rs
@@ -1,5 +1,7 @@
 #[macro_use]
 extern crate diesel;
 
+pub mod s3;
 pub mod scraper;
 pub mod sql;
+pub mod error;
