--- atcoder-problems-dumper-lambda/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "atcoder-problems-dumper-lambda"
-version = "0.1.0"
-authors = ["Kenko Nakamura <kenkou.n@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-atcoder-problems-sql-common = { path = "../atcoder-problems-sql-common" }
-lambda_runtime = "^0.1"
-diesel = { version = "1.4.2", features = ["postgres"] }
-openssl-probe = "0.1.2"
-openssl = "*"
-log = "^0.4"
-simple_logger = "^1"
-rusoto_s3 = "0.37.0"
-rusoto_core = "0.37.0"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-
-[[bin]]
-name = "bootstrap"
-path = "src/main.rs"
