--- atcoder-problems-dumper-lambda/src/lib.rs
@@ -1,5 +1,0 @@
-#[macro_use]
-extern crate diesel;
-
-pub mod models;
-pub mod s3;
