--- atcoder-problems-dumper-lambda/src/main.rs
@@ -1,153 +1,0 @@
-#[macro_use]
-extern crate log;
-extern crate simple_logger;
-
-#[macro_use]
-extern crate lambda_runtime as lambda;
-extern crate openssl;
-
-use atcoder_problems_dumper_lambda::models::*;
-use atcoder_problems_dumper_lambda::s3::S3Client;
-use atcoder_problems_sql_common::schema::*;
-use diesel::prelude::*;
-use diesel::{sql_query, Connection, PgConnection};
-use lambda::error::HandlerError;
-use openssl_probe;
-use std::env;
-use std::error::Error;
-
-fn main() -> Result<(), Box<dyn Error>> {
-    simple_logger::init_with_level(log::Level::Info)?;
-    openssl_probe::init_ssl_cert_env_vars();
-    lambda!(my_handler);
-
-    Ok(())
-}
-
-fn my_handler(_: String, c: lambda::Context) -> Result<String, HandlerError> {
-    info!("Started!");
-    let url = env::var("SQL_URL").new_err(&c)?;
-    let conn = PgConnection::establish(&url).new_err(&c)?;
-
-    let merged_query = sql_query(r"
-            SELECT
-                problems.id,
-                problems.contest_id,
-                problems.title,
-                shortest.submission_id AS shortest_submission_id,
-                shortest.problem_id AS shortest_problem_id,
-                shortest.contest_id AS shortest_contest_id,
-                shortest_submissions.user_id AS shortest_user_id,
-                fastest.submission_id AS fastest_submission_id,
-                fastest.problem_id AS fastest_problem_id,
-                fastest.contest_id AS fastest_contest_id,
-                fastest_submissions.user_id AS fastest_user_id,
-                first.submission_id AS first_submission_id,
-                first.problem_id AS first_problem_id,
-                first.contest_id AS first_contest_id,
-                first_submissions.user_id AS first_user_id,
-                shortest_submissions.length AS source_code_length,
-                fastest_submissions.execution_time AS execution_time,
-                points.point,
-                points.predict,
-                solver.user_count AS solver_count
-            FROM
-                problems
-                LEFT JOIN shortest ON shortest.problem_id = problems.id
-                LEFT JOIN fastest ON fastest.problem_id = problems.id
-                LEFT JOIN first ON first.problem_id = problems.id
-                LEFT JOIN submissions AS shortest_submissions ON shortest.submission_id = shortest_submissions.id
-                LEFT JOIN submissions AS fastest_submissions ON fastest.submission_id = fastest_submissions.id
-                LEFT JOIN submissions AS first_submissions ON first.submission_id = first_submissions.id
-                LEFT JOIN points ON points.problem_id = problems.id
-                LEFT JOIN solver ON solver.problem_id = problems.id
-                ORDER BY problems.id;
-        ");
-
-    let client = S3Client::new();
-
-    client
-        .update(
-            merged_query.load::<MergedProblem>(&conn).new_err(&c)?,
-            "resources/merged-problems.json",
-        )
-        .new_err(&c)?;
-    client
-        .update(
-            contests::table
-                .order_by(contests::id)
-                .load::<Contest>(&conn)
-                .new_err(&c)?,
-            "resources/contests.json",
-        )
-        .new_err(&c)?;
-    client
-        .update(
-            accepted_count::table
-                .order_by(accepted_count::user_id)
-                .load::<UserProblemCount>(&conn)
-                .new_err(&c)?,
-            "resources/ac.json",
-        )
-        .new_err(&c)?;
-    client
-        .update(
-            problems::table
-                .order_by(problems::id)
-                .load::<Problem>(&conn)
-                .new_err(&c)?,
-            "resources/problems.json",
-        )
-        .new_err(&c)?;
-    client
-        .update(
-            rated_point_sum::table
-                .order_by(rated_point_sum::user_id)
-                .load::<UserSum>(&conn)
-                .new_err(&c)?,
-            "resources/sums.json",
-        )
-        .new_err(&c)?;
-    client
-        .update(
-            language_count::table
-                .order_by(language_count::user_id)
-                .load::<UserLanguageCount>(&conn)
-                .new_err(&c)?,
-            "resources/lang.json",
-        )
-        .new_err(&c)?;
-    client
-        .update(
-            minimum_performances::table
-                .order_by(minimum_performances::problem_id)
-                .load::<MinimumPerformance>(&conn)
-                .new_err(&c)?,
-            "resources/problem-performances.json",
-        )
-        .new_err(&c)?;
-    client
-        .update(
-            contest_problem::table
-                .order_by(contest_problem::problem_id)
-                .load::<ContestProblem>(&conn)
-                .new_err(&c)?,
-            "resources/contest-problem.json",
-        )
-        .new_err(&c)?;
-
-    Ok(String::new())
-}
-
-trait ErrorMapper<S> {
-    fn new_err(self, c: &lambda::Context) -> Result<S, HandlerError>;
-}
-
-impl<S, T> ErrorMapper<S> for Result<S, T>
-where
-    T: std::fmt::Debug,
-{
-    fn new_err(self, c: &lambda::Context) -> Result<S, HandlerError> {
-        self.map_err(|e| c.new_error(&format!("{:?}", e)))
-    }
-}
