--- atcoder-problems-monitor/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "atcoder-problems-monitor"
-version = "0.1.0"
-authors = ["kenkoooo <kenkou.n@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-rusoto_cloudwatch = "0.39.0"
-rusoto_s3 = "0.39.0"
-rusoto_core = "0.39.0"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-base64 = "0.10"
-lambda_runtime = "0.2"
-simple_logger = "1.2.0"
-log = "0.4.6"
-openssl-probe = "0.1.2"
-openssl = "*"
